api-gateway/requirements.txt
django>=4.2,<5.0
requests>=2.31
PyJWT>=2.8
